# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "articleManger",
        "contact": {
            "name": "mylxsw",
            "email": "mylxsw@aicode.cc"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "開發環境"
        }
    ],
    "paths": {
...
        "/translation/index": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Get translations by article ID and language",
                "operationId": "2d258c3dc120a0e47a767bb9922c935d",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "description": "文章標題",
                                                "type": "string"
                                            },
                                            "content": {
                                                "description": "文章內容",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/translation/create": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "建立新文章",
                "operationId": "6ecbc8b06cb5d854670fc5b591151417",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "article_id": {
                                        "description": "Article ID",
                                        "type": "integer"
                                    },
                                    "article": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "description": "文章標題",
                                                    "type": "string"
                                                },
                                                "content": {
                                                    "description": "文章內容",
                                                    "type": "string"
                                                },
                                                "language": {
                                                    "description": "語系",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "創建完成",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "語言已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/translation/{id}": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Get translation by ID",
                "operationId": "1944ddb8256a96373a038d590ff2ebda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Translation ID",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Translation title",
                                            "type": "string"
                                        },
                                        "content": {
                                            "description": "Translation content",
                                            "type": "string"
                                        },
                                        "language": {
                                            "description": "Translation language",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                }
            }
        },
        "/translation/update/{id}": {
            "put": {
                "tags": [
                    "Translation"
                ],
                "summary": "update Translation",
                "operationId": "6e7befaecab7f8271cc8abb3595d88a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "article_id": {
                                        "description": "Article ID",
                                        "type": "integer"
                                    },
                                    "title": {
                                        "description": "Translation title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Translation content",
                                        "type": "string"
                                    },
                                    "language": {
                                        "description": "Translation language",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新完成",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "語言已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "查無此資料"
                    }
                }
            }
        },
        "/translation/delete/{id}": {
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "刪除文章",
                "operationId": "3ecaced55fddbf7ab516e4b8b8963631",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除完成",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}